backend: feat: add proper error handling to models and adjust controllers

+++ motionLab_app/backend/controllers/BVH_controller.py
@@ -10,14 +10,18 @@ class BVHController:
         project_id = data["projectId"]
         
         bvh = BVH.create(name, path, project_id)
-        print("BVH created:", bvh)
+        if not bvh:
+            return {"success": False, "error": "Error creating BVH"}
+        
         return {"success": True, "bvh": bvh.to_dict()}
     
     @staticmethod
     def create_bvhs(filenames, project_id):
         try:
             for filename in filenames:
-                BVH.create(filename, project_id)
+                bvh = BVH.create(filename, project_id)
+                if not bvh:
+                    return False
             
             return True
         except Exception as e:
@@ -27,6 +31,9 @@ class BVHController:
     def delete_bvhs_by_project_id(project_id):
         try:
             bvh_filenames = BVH.delete_bvhs_by_project_id(project_id)
+            if not bvh_filenames:
+                return False
+            
             for filename in bvh_filenames:
                 os.remove(os.path.join("BVHs", filename))
                 
@@ -36,4 +43,8 @@ class BVHController:
         
     @staticmethod
     def get_bvhs_by_project_id(project_id):
-        return BVH.get_by_project_id(project_id)
+        bvhs = BVH.get_by_project_id(project_id)
+        if not bvhs:
+            return None
+        
+        return [bvh.to_dict() for bvh in bvhs]
